requirements.txt
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.6.0
amqp==2.5.0
aniso8601==6.0.0
APScheduler==3.6.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
crcmod==1.7
decorator==4.4.0
Flask==1.0.2
Flask-APScheduler==1.11.0
Flask-JWT-Extended==3.18.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
lxml==4.3.3
MarkupSafe==1.1.1
mongoengine==0.17.0
mysql-connector==2.2.9
nltk==3.4.1
oss2==2.6.1
pbr==5.2.0
pdfminer2==20151206
pdfx==1.3.0
peewee==3.9.5
pycryptodome==3.8.1
PyJWT==1.7.1
pymongo==3.8.0
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
tzlocal==1.5.1
urllib3==1.24.3
vine==1.3.0
Werkzeug==0.15.2
WTForms==2.2.1
